chore: Merge rest of derive_codec.rs and json.rs
